pyproject.toml
[tool.poetry]
name = "content-analytics-pipeline"
version = "0.1.0"
description = "A pipeline for processing media analytics events"
packages = [
    { include = "content_analytics" }
]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
kafka-python-ng = "^2.2.3"
pandas = "^2.2.3"
pyarrow = "^19.0.1"
pydantic-settings = "^2.8.1"
redis = "^5.2.1"
celery = "^5.4.0"
fastparquet = "^2024.11.0"
boto3 = "^1.37.20"
six = "^1.17.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"




[tool.poetry.group.dev.dependencies]
ruff = "^0.11.2"
kafka-python-ng = "^2.2.3"
faker = "^37.1.0"
pydantic-settings = "^2.8.1"
pytest = "^8.3.5"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"